disease_diagnosis_backend/requirements.txt
Django==4.1.7
djangorestframework==3.14.0
django-cors-headers==3.14.0
pandas==1.5.3
gunicorn==20.1.0
psycopg2-binary==2.9.5
